chore: Updated prefixed select component
