+++ app/helpers/view_helper.rb
@@ -20,9 +20,9 @@ module ViewHelper
     input_group: 'Viral::Form::InputGroupComponent',
     pageheader: 'Viral::PageHeaderComponent',
     prefixed_checkbox: 'Viral::Form::Prefixed::CheckboxComponent',
+    prefixed_select: 'Viral::Form::Prefixed::SelectComponent',
     pill: 'Viral::PillComponent',
     select: 'Viral::Form::SelectComponent',
-    select_group: 'Viral::Form::SelectGroupComponent',
     tabs: 'Viral::TabsComponent',
     text_input: 'Viral::Form::TextInputComponent',
     time_ago: 'Viral::TimeAgoComponent',
